break(Universal): switch to a more intuitive API / more conventional behavior
