+++ src/universal/universal.ts
@@ -4,38 +4,17 @@ export {
 
 import type { any } from "../any/exports.js"
 
-
 declare namespace Universal {
-  type parseNumeric<type> = type extends `${infer x extends number}` ? x : never
+  /** @internal */
+  type parseInt<t> = [t] extends [`${infer n extends number}`] ? n : t;
 
-  type key<key extends any.index> =
-    | `${Exclude<key, symbol>}`
-    | parseNumeric<key>
-    | key
-    ;
+  /** ### {@link Universal.get `Universal.get`} */
+  type get<t, k extends any.index> = t[keyOf<t, k>]
 
-  type keyof<type>
-    = type extends any.array
-    ? Universal.key<Extract<keyof type, `${number}`>>
-    : Universal.key<keyof type>
-    ;
+  /** ### {@link Universal.keyOf `Universal.keyOf`} */
+  type keyOf<t, k extends any.index = keyof t>
+    = (k extends any.key ? (`${k}` | parseInt<k>) : unknown) & keyof t
 
-  type get<index extends any.key, type>
-    = index extends keyof type ? type[index]
-    : parseNumeric<index> extends any.number<infer x>
-    ? x extends keyof type ? type[x]
-    : never
-    : `${index}` extends any.string<infer s>
-    ? s extends keyof type ? type[s]
-    : never
-    : never
-    ;
-
-  type values<type>
-    = type extends any.array
-    ? type[number]
-    : type extends any.object
-    ? type[keyof type]
-    : type
-    ;
+  /** ### {@link Universal.key `Universal.key`} */
+  type key<k> = `${k & any.showable}` | globalThis.Exclude<k, any.showable>
 }
